refactor: extract group auto assignment
