+++ src/orpc/groups/dal.ts
@@ -1,8 +1,11 @@
 import {
+  autoAssignGroup,
+  autoAssignAllEligible as runAutoAssignAllEligible,
+} from './auto-assign';
+import {
   assignAthleteToGroup,
   unassignAthleteFromGroup,
 } from './assign-athlete';
-import type { Prisma } from '@/generated/prisma/client';
 import type {
   AssignAthleteDTO,
   AutoAssignDTO,
@@ -10,15 +13,10 @@ import type {
   UnassignAthleteDTO,
   UpdateGroupDTO,
 } from './dto';
-import { recordTournamentActivity } from '@/orpc/activity/dal';
 import { prisma } from '@/lib/db';
 import { publishSelectionInvalidate } from '@/lib/tournament/tournament-realtime-broadcast';
 
 export class GroupDAL {
-  private static readonly UNASSIGNED_GROUP_FILTER = {
-    groupId: null,
-  } satisfies Prisma.TournamentAthleteWhereInput;
-
   static async findByTournamentId(tournamentId: string) {
     return prisma.group.findMany({
       where: { tournamentId },
@@ -70,86 +68,14 @@ export class GroupDAL {
   }
 
   static async autoAssign(input: AutoAssignDTO & { adminId: string }) {
-    const result = await prisma.$transaction(async (tx) => {
-      const group = await tx.group.findUnique({
-        where: { id: input.groupId },
-      });
-      if (!group) throw new Error('Group not found');
-
-      const where: Prisma.TournamentAthleteWhereInput = {
-        tournamentId: input.tournamentId,
-        AND: [GroupDAL.UNASSIGNED_GROUP_FILTER],
-      };
-
-      if (group.gender) where.gender = group.gender;
-      if (group.beltMin != null || group.beltMax != null) {
-        where.beltLevel = {
-          ...(group.beltMin != null ? { gte: group.beltMin } : {}),
-          ...(group.beltMax != null ? { lte: group.beltMax } : {}),
-        };
-      }
-      if (group.weightMin != null || group.weightMax != null) {
-        where.weight = {
-          ...(group.weightMin != null ? { gte: group.weightMin } : {}),
-          ...(group.weightMax != null ? { lte: group.weightMax } : {}),
-        };
-      }
-
-      const unassigned = await tx.tournamentAthlete.findMany({ where });
-
-      if (unassigned.length === 0) return { assigned: 0 };
-
-      await tx.tournamentAthlete.updateMany({
-        where: { id: { in: unassigned.map((a) => a.id) } },
-        data: { groupId: input.groupId, status: 'assigned' },
-      });
-
-      await recordTournamentActivity(
-        {
-          tournamentId: input.tournamentId,
-          adminId: input.adminId,
-          eventType: 'group.auto_assign',
-          entityType: 'group',
-          entityId: input.groupId,
-          payload: { count: unassigned.length },
-        },
-        tx
-      );
-
-      return { assigned: unassigned.length };
-    });
-    publishSelectionInvalidate(input.tournamentId);
-    return result;
+    return autoAssignGroup(input);
   }
 
   static async autoAssignAllEligible(input: {
     tournamentId: string;
     adminId: string;
   }) {
-    const groups = await prisma.group.findMany({
-      where: { tournamentId: input.tournamentId },
-      include: { _count: { select: { matches: true } } },
-    });
-
-    let assigned = 0;
-    let groupsRun = 0;
-    let groupsSkipped = 0;
-
-    for (const group of groups) {
-      if (group._count.matches > 0) {
-        groupsSkipped += 1;
-        continue;
-      }
-      groupsRun += 1;
-      const result = await GroupDAL.autoAssign({
-        tournamentId: input.tournamentId,
-        groupId: group.id,
-        adminId: input.adminId,
-      });
-      assigned += result.assigned;
-    }
-
-    return { assigned, groupsRun, groupsSkipped };
+    return runAutoAssignAllEligible(input);
   }
 
   static async assignAthlete(input: AssignAthleteDTO & { adminId: string }) {
